storage/provider: abstract "df" usage
